+++ Notes/reference/releaseNotes.md
@@ -3,6 +3,11 @@
 Some basic information on our tagged releases.
 Note - `git log tag1..tag2` shows the commit log between versions.
 
+## Version 1.0.5
+
+* bug fix in lambdaBridge handling of SameSite cookie
+* bug fix in lambdaBridge handling of redirect URLs
+
 ## Version 1.0.4
 
 * bug fix in lambdaBridge handling of origin for CORS